Fix more clippy-isms, enable clippy in CI and Makefile

+++ soroban-env-host/benches/common/cost_types/val_xdr_conv.rs
@@ -11,7 +11,7 @@ impl HostCostMeasurement for ValXdrConvMeasure {
     fn new_random_case(_host: &Host, rng: &mut StdRng, _input: u64) -> (Option<RawVal>, ScVal) {
         let v = rng.next_u32();
         let rv: RawVal = v.into();
-        let scv: ScVal = ScVal::U32(v as u32);
+        let scv: ScVal = ScVal::U32(v);
         (Some(rv), scv)
     }
 }
